requirements.txt
fastapi
uvicorn
sentence-transformers
python-multipart
mysql-connector-python
python-dotenv
requests
pdfplumber
paddleocr
tqdm
matplotlib
reportlab
